set transaction id type to bigint

+++ src/main/java/ru/portfolio/portfolio/entity/TransactionCashFlowEntityPK.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
 @Data
 public class TransactionCashFlowEntityPK implements Serializable {
     @Column(name = "transaction_id")
-    private int transactionId;
+    private long transactionId;
 
     @Column(name = "type")
     private int type;
